bpf: create a header for struct bpf_link
struct bpf_link needs to be embedded by cgroups.
Put it in its own header.
